// types/purchase.ts
export interface Material {
  id: number;
  name: string;
  type: string;
  type_display: string;
  available: string;
  code_name: string;
}

export interface EachPurchaseMaterial {
  id: number;
  material: Material;
  amount: string;
  price: string;
  total_price: string;
  date: string;
}

export interface User {
  id: number;
  telegram_id: number;
  telegram_user_name: string;
  role: Array<{
    id: number;
    Name: string;
    date: string;
  }>;
  first_name: string;
}

export interface Purchase {
  id: number;
  created_by: User;
  confirmed_by:User;
  each_material_purchase: EachPurchaseMaterial[];
  from_account: any;
  total_amount: string;
  to_account_number: string;
  to_account_name: string;
  invoice: boolean;
  invoice_image: string | null;
  status: string;
  request_status: string;
  payment_screenshot: string | null;
  payment_code: string | null;
  is_deleted: boolean;
  date: string;
  from_wallet: string;
}

export interface PurchaseListResponse {
  count: number;
  next: string | null;
  previous: string | null;
  results: Purchase[];
}

export interface Account {
  id: number;
  bank: string;
  available_amount: string;
  account_number: string;
  account_type: string;
  account_name: string;
  deleted: boolean;
  date: string;
}